requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-filter==2.1.0
django-mathfilters==0.4.0
djangorestframework==3.9.4
Markdown==3.1.1
mysqlclient==1.4.2.post1
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
